requirements.txt
selenium==4.34.2
webdriver-manager==4.0.2
requests==2.32.4
pytest==8.0.0
pytest-cov==5.0.0
sqlite3
python-docx==1.2.0
pypdf==4.2.0
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
python-dotenv==1.1.1 